+++ Octokit/Clients/INotificationsClient.cs
@@ -21,10 +21,101 @@ namespace Octokit
         Task<IReadOnlyList<Notification>> GetAllForCurrent();
 
         /// <summary>
+        /// Retrieves all of the <see cref="Notification"/>s for the current user.
+        /// </summary>
+        /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
+        /// <returns>A <see cref="IReadOnlyPagedCollection{Notification}"/> of <see cref="Notification"/>.</returns>
+        [SuppressMessage("Microsoft.Design", "CA1024:UsePropertiesWhereAppropriate")]
+        Task<IReadOnlyList<Notification>> GetAllForCurrent(NotificationsRequest request);
+
+        /// <summary>
         /// Retrieves all of the <see cref="Notification"/>s for the current user specific to the specified repository.
         /// </summary>
         /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
         /// <returns>A <see cref="IReadOnlyPagedCollection{Notification}"/> of <see cref="Notification"/>.</returns>
         Task<IReadOnlyList<Notification>> GetAllForRepository(string owner, string name);
+
+        /// <summary>
+        /// Retrieves all of the <see cref="Notification"/>s for the current user specific to the specified repository.
+        /// </summary>
+        /// <exception cref="AuthorizationException">Thrown if the client is not authenticated.</exception>
+        /// <returns>A <see cref="IReadOnlyPagedCollection{Notification}"/> of <see cref="Notification"/>.</returns>
+        Task<IReadOnlyList<Notification>> GetAllForRepository(string owner, string name, NotificationsRequest request);
+
+        /// <summary>
+        /// Marks all notifications as read.
+        /// </summary>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-as-read</remarks>
+        /// <returns></returns>
+        Task MarkAsRead();
+
+        /// <summary>
+        /// Marks all notifications as read.
+        /// </summary>
+        /// <param name="markAsReadRequest">The <see cref="MarkAsReadRequest"/> parameter which specifies which notifications to mark.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-as-read</remarks>
+        /// <returns></returns>
+        Task MarkAsRead(MarkAsReadRequest markAsReadRequest);
+
+        /// <summary>
+        /// Marks the notifications for a given repository as read.
+        /// </summary>
+        /// <param name="owner">The owner of the repository</param>
+        /// <param name="name">The name of the repository</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read-in-a-repository</remarks>
+        /// <returns></returns>
+        Task MarkAsReadForRepository(string owner, string name);
+
+        /// <summary>
+        /// Marks the notifications for a given repository as read.
+        /// </summary>
+        /// <param name="owner">The owner of the repository</param>
+        /// <param name="name">The name of the repository</param>
+        /// <param name="markAsRead">The <see cref="MarkAsReadRequest"/> parameter which specifies which notifications to mark.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-notifications-as-read-in-a-repository</remarks>
+        /// <returns></returns>
+        Task MarkAsReadForRepository(string owner, string name, MarkAsReadRequest markAsRead);
+
+        /// <summary>
+        /// Retrives a single <see cref="Notification"/> by Id.
+        /// </summary>
+        /// <param name="id">The Id of the notification to retrieve.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#view-a-single-thread</remarks>
+        /// <returns>A <see cref="Notification"/> for the given Id.</returns>
+        [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get")]
+        Task<Notification> Get(int id);
+
+        /// <summary>
+        /// Marks a single notification as read.
+        /// </summary>
+        /// <param name="id">The id of the notification.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#mark-a-thread-as-read</remarks>
+        /// <returns></returns>
+        Task MarkAsRead(int id);
+
+        /// <summary>
+        /// Retrives a <see cref="ThreadSubscription"/> for the provided thread id.
+        /// </summary>
+        /// <param name="id">The Id of the thread to retrieve subscription status.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#get-a-thread-subscription</remarks>
+        /// <returns>A <see cref="ThreadSubscription"/> for the chosen thread.</returns>
+        Task<ThreadSubscription> GetThreadSubscription(int id);
+
+        /// <summary>
+        /// Sets the authenticated user's subscription settings for a given thread.
+        /// </summary>
+        /// <param name="id">The Id of the thread to update.</param>
+        /// <param name="threadSubscription">The subscription parameters to set.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#set-a-thread-subscription</remarks>
+        /// <returns></returns>
+        Task<ThreadSubscription> SetThreadSubscription(int id, NewThreadSubscription threadSubscription);
+
+        /// <summary>
+        /// Deletes the authenticated user's subscription to a given thread.
+        /// </summary>
+        /// <param name="id">The Id of the thread to delete subscription from.</param>
+        /// <remarks>http://developer.github.com/v3/activity/notifications/#delete-a-thread-subscription</remarks>
+        /// <returns></returns>
+        Task DeleteThreadSubscription(int id);
     }
 }