+++ Octokit/Clients/RepositoryCommitsClient.cs
@@ -3,6 +3,12 @@ using System.Threading.Tasks;
 
 namespace Octokit
 {
+    /// <summary>
+    /// A client for GitHub's Repository Commits API.
+    /// </summary>
+    /// <remarks>
+    /// See the <a href="http://developer.github.com/v3/repos/commits/">Repository Commits API documentation</a> for more information.
+    /// </remarks>
     public class RepositoryCommitsClient : IRepositoryCommitsClient
     {
         readonly IApiConnection _apiConnection;