+++ Octokit/Models/Request/MilestoneUpdate.cs
@@ -8,11 +8,6 @@ namespace Octokit
     public class MilestoneUpdate
     {
         /// <summary>
-        /// The milestone number.
-        /// </summary>
-        public int Number { get; set; }
-
-        /// <summary>
         /// Title of the milestone (required)
         /// </summary>
         public string Title { get; set; }
@@ -20,7 +15,7 @@ namespace Octokit
         /// <summary>
         /// Whether the milestone is open or closed. The default is <see cref="ItemState.Open"/>.
         /// </summary>
-        public ItemState State { get; set; }
+        public ItemState? State { get; set; }
 
         /// <summary>
         /// Optional description for the milestone.